Move those methods back
