+++ src/main/java/com/zyurkalov/ideavim/syntaxtreejumper/motions/AbstractFindNodeMotionHandler.java
@@ -9,13 +9,15 @@ import org.jetbrains.annotations.NotNull;
 import java.util.Optional;
 import java.util.function.Function;
 
+import static com.zyurkalov.ideavim.syntaxtreejumper.adapters.SyntaxTreeAdapter.getChild;
+
 public abstract class AbstractFindNodeMotionHandler implements MotionHandler {
     protected final SyntaxTreeAdapter syntaxTree;
     protected final MotionDirection direction;
-    private final SyntaxTreeAdapter.WhileSearching whileSearching;
+    private final WhileSearching whileSearching;
 
     public AbstractFindNodeMotionHandler(SyntaxTreeAdapter syntaxTree, MotionDirection direction,
-                                         SyntaxTreeAdapter.WhileSearching whileSearching) {
+                                         WhileSearching whileSearching) {
         this.syntaxTree = syntaxTree;
         this.direction = direction;
         this.whileSearching = whileSearching;
@@ -32,9 +34,9 @@ public abstract class AbstractFindNodeMotionHandler implements MotionHandler {
         }
 
         // 2. Search for nodes based on a direction and searching criteria
-        Optional<SyntaxNode> targetListNode = syntaxTree.findNodeByDirection(
+        Optional<SyntaxNode> targetListNode = findNodeByDirection(
                 currentNode, direction, initialOffsets,
-                createFunctionToCheckSearchingCriteria(direction, initialOffsets, whileSearching), 
+                createFunctionToCheckSearchingCriteria(direction, initialOffsets, whileSearching),
                 whileSearching);
         if (targetListNode.isPresent()) {
             // 3. Place caret at the first child
@@ -51,5 +53,81 @@ public abstract class AbstractFindNodeMotionHandler implements MotionHandler {
     }
 
     @NotNull
-    public abstract Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(MotionDirection direction, Offsets initialSelection, SyntaxTreeAdapter.WhileSearching whileSearching);
+    public abstract Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(
+            MotionDirection direction, Offsets initialSelection, WhileSearching whileSearching);
+
+    private static boolean isNodeReallyFound(MotionDirection direction, Offsets initialSelection, Optional<SyntaxNode> found) {
+        return found.isPresent() &&
+                found.get().isInRightDirection(initialSelection, direction) &&
+                !found.get().areBordersEqual(initialSelection);
+    }
+
+    /**
+     * Finds a PARAMETER_LIST or ARGUMENT_LIST node based on the direction from the current node.
+     *
+     * @param currentNode                   The starting node
+     * @param direction                     The direction to search
+     * @param initialSelection              The initial selection
+     * @param functionToCheckSearchCriteria The function to find parameter nodes based on specific criteria
+     * @return The found parameter/argument list node or null if not found
+     */
+    public Optional<SyntaxNode> findNodeByDirection(
+            SyntaxNode currentNode, MotionDirection direction, Offsets initialSelection,
+            @NotNull Function<SyntaxNode, Optional<SyntaxNode>> functionToCheckSearchCriteria, WhileSearching whileSearching) {
+        Optional<SyntaxNode> found = findWithinNeighbours(
+                currentNode, direction, initialSelection, functionToCheckSearchCriteria, whileSearching);
+
+        while (found.isEmpty()) {
+            currentNode = currentNode.getParent();
+            if (currentNode == null || currentNode.isPsiFile()) {
+                break;
+            }
+            found = findWithinNeighbours(currentNode, direction, initialSelection, functionToCheckSearchCriteria, whileSearching);
+        }
+        return found;
+    }
+
+    public enum WhileSearching {
+        SKIP_INITIAL_SELECTION, DO_NOT_SKIP_INITIAL_SELECTION
+    }
+
+    public Optional<SyntaxNode> findWithinNeighbours(
+            @NotNull SyntaxNode currentNode, MotionDirection direction, Offsets initialSelection,
+            Function<SyntaxNode, Optional<SyntaxNode>> findNodeType,
+            WhileSearching whileSearching) {
+        Optional<SyntaxNode> found = Optional.empty();
+        var next = Optional.of(currentNode);
+        while (next.isPresent()) {
+            currentNode = next.get();
+            if (whileSearching == WhileSearching.SKIP_INITIAL_SELECTION && !currentNode.isInRightDirection(initialSelection, direction)) {
+                next = nextNeighbour(currentNode, direction); //TODO: I think we need to move it here as a virtual method
+                continue;
+            }
+            found = findNodeType.apply(currentNode);
+            if (isNodeReallyFound(direction, initialSelection, found)
+            ) {
+                return found;
+            }
+            if (!currentNode.getChildren().isEmpty()) {
+                found = findWithinNeighbours(
+                        getChild(currentNode, direction), direction, initialSelection, findNodeType, WhileSearching.DO_NOT_SKIP_INITIAL_SELECTION);
+                if (isNodeReallyFound(direction, initialSelection, found)
+                ) {
+                    return found;
+                }
+            }
+            next = nextNeighbour(currentNode, direction);
+        }
+        return found;
+    }
+
+    // TODO: rename this method and combine with get child and getParent
+    public Optional<SyntaxNode> nextNeighbour(SyntaxNode node, MotionDirection direction) {
+        return switch (direction) {
+            case FORWARD -> Optional.ofNullable(node.getNextSibling());
+            case BACKWARD -> Optional.ofNullable(node.getPreviousSibling());
+            case EXPAND -> Optional.ofNullable(node.getParent());
+            case SHRINK -> Optional.ofNullable(node.getFirstChild());
+        };
+    }
 }
