+++ src/main/java/com/zyurkalov/ideavim/syntaxtreejumper/motions/ArgumentParameterListMotionHandler.java
@@ -17,11 +17,11 @@ import java.util.function.Function;
 public class ArgumentParameterListMotionHandler extends AbstractFindNodeMotionHandler {
 
     public ArgumentParameterListMotionHandler(SyntaxTreeAdapter syntaxTree, MotionDirection direction) {
-        super(syntaxTree, direction, SyntaxTreeAdapter.WhileSearching.DO_NOT_SKIP_INITIAL_SELECTION);
+        super(syntaxTree, direction, WhileSearching.DO_NOT_SKIP_INITIAL_SELECTION);
     }
     @Override
     @NotNull
-    public Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(MotionDirection direction, Offsets initialSelection, SyntaxTreeAdapter.WhileSearching whileSearching) {
+    public Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(MotionDirection direction, Offsets initialSelection, WhileSearching whileSearching) {
         return node -> {
             if (node.isFunctionParameter() || node.isFunctionArgument() || node.isTypeParameter()) {
                 return Optional.of(node);
