+++ src/main/java/com/zyurkalov/ideavim/syntaxtreejumper/motions/MethodFunctionMotionHandler.java
@@ -17,12 +17,12 @@ import java.util.function.Function;
 public class MethodFunctionMotionHandler extends AbstractFindNodeMotionHandler {
 
     public MethodFunctionMotionHandler(SyntaxTreeAdapter syntaxTree, MotionDirection direction) {
-        super(syntaxTree, direction, SyntaxTreeAdapter.WhileSearching.DO_NOT_SKIP_INITIAL_SELECTION);
+        super(syntaxTree, direction, WhileSearching.DO_NOT_SKIP_INITIAL_SELECTION);
     }
 
     @Override
     @NotNull
-    public Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(MotionDirection direction, Offsets initialSelection, SyntaxTreeAdapter.WhileSearching whileSearching) {
+    public Function<SyntaxNode, Optional<SyntaxNode>> createFunctionToCheckSearchingCriteria(MotionDirection direction, Offsets initialSelection, WhileSearching whileSearching) {
         return node -> {
             if (node.isMethodDefinition() || node.isFunctionDefinition()) {
                 return Optional.of(node);
